chg: remove unused question_type xls + code
- the .xls files in the question_types directory haven't changed since
  2014 and since then new question types have just been added to
  QUESTION_TYPE_DICT and documented externally.
- deleted associated unused code from question_type_dictionary.py and
  xls2json.py which was mostly a duplication of existing workbook
  reading code.
- moved print_pyobj_to_json to utils.py since it doesn't relate to
  anything specific to xls2json.py and it so simplifies module imports
