+++ pyxform/question_type_dictionary.py
@@ -4,20 +4,6 @@ XForm survey question type mapping dictionary module.
 
 from types import MappingProxyType
 
-from pyxform.xls2json import QuestionTypesReader, print_pyobj_to_json
-
-
-def generate_new_dict():
-    """
-    This is just here incase there is ever any need to generate the question
-    type dictionary from all.xls again.
-    It shouldn't be called as part of any application.
-    """
-    path_to_question_types = "/pyxform/question_types/all.xls"
-    json_dict = QuestionTypesReader(path_to_question_types).to_json_dict()
-    print_pyobj_to_json(json_dict, "new_question_type_dict.json")
-
-
 _QUESTION_TYPE_DICT = {
     "q picture": {
         "control": {"tag": "upload", "mediatype": "image/*"},
